Restructure packages to follow Clean Architecture Created separate Use-cases from MainViewModel Recode MainViewModel.kt
