+++ app/build.gradle
@@ -63,7 +63,9 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.5.0'
     implementation "com.google.accompanist:accompanist-systemuicontroller:0.23.1"
     implementation "com.google.accompanist:accompanist-permissions:0.23.1"
-
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.9'
+    implementation "com.squareup.retrofit2:adapter-rxjava2:2.5.0"
     implementation "com.google.dagger:hilt-android:$hilt_version"
     kapt "com.google.dagger:hilt-compiler:$hilt_version"
     implementation 'com.google.android.gms:play-services-location:20.0.0'
